+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [3.3.2] - 2023-03-29
+### Fixed
+- toQuantity() now works with subclasses of Quantity.
+
 ## [3.3.1] - 2022-11-22
 ### Fixed
 - Aggregate init parameter can be any type, not just an integer.